Lineage Tree Refactor
